+++ Sources/Sentry/SentryClient.m
@@ -168,12 +168,14 @@ NSString *const DropSessionLogMessage = @"Session has no release name. Won't sen
 - (SentryEvent *)buildErrorEvent:(NSError *)error
 {
     SentryEvent *event = [[SentryEvent alloc] initWithLevel:kSentryLevelError];
-    NSString *formatted = [NSString stringWithFormat:@"%@ %ld", error.domain, (long)error.code];
-    SentryMessage *message = [[SentryMessage alloc] initWithFormatted:formatted];
-    message.message = [error.domain stringByAppendingString:@" %s"];
-    message.params = @[ [NSString stringWithFormat:@"%ld", (long)error.code] ];
-    event.message = message;
+    NSString *errorCodeAsString = [NSString stringWithFormat:@"%ld", (long)error.code];
+    SentryException *sentryException = [[SentryException alloc] initWithValue:errorCodeAsString
+                                                                         type:error.domain];
+    event.exceptions = @[ sentryException ];
+    event.fingerprint = @[ error.domain, [NSString stringWithFormat:@"%ld", (long)error.code] ];
+
     [self setUserInfo:error.userInfo withEvent:event];
+
     return event;
 }